Swagger Funcionando al 100% - 23:30

+++ src/main/java/com/Perfulandia/perfulandia/model/Producto.java
@@ -1,5 +1,7 @@
 package com.Perfulandia.perfulandia.model;
 
+import java.math.BigDecimal;
+
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -9,6 +11,7 @@ import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
 
 @Entity
 @Table(name = "productos")
@@ -31,5 +34,5 @@ public class Producto {
     private Integer stock;
 
     @Column(nullable = false)
-    private Double precio;
+    private BigDecimal precio;
 }
